added the last element of page, edited some elements of the middle input box, just need to add the functions in now
