+++ src/App.js
@@ -1,11 +1,13 @@
 import Nav from './components/Nav';
 import InputBox from './components/InputBox';
+import SelectBox from './components/SelectBox';
 
 function App() {
   return (
     <div className='App'>
       <Nav />
       <InputBox />
+      <SelectBox />
       
     </div>
   )
