+++ src/components/selectBox.js
@@ -1,18 +1,16 @@
-import Select from 'react-select'
+import './styles/selectbox.css';
+import { TextField } from "@mui/material"
 
-const options = ([
-    { value: 'decimal', label: 'Decimal' },
-    { value: 'binary', label: 'Binary' },
-    { value: 'hexadecimal', label: 'Hexadecimal' },
-    { value: 'octal', label: 'Octal' }
-  ]);
-
-  
-const selectBox = () => {
+function SelectBox () {
     return (
-        <Select options={ option }
-        />
+        <div className="main-box">
+            <TextField className='output-box' label="Decimal" InputProps={{ readOnly: true }} />
+            <TextField className='output-box' label="Binary" InputProps={{ readOnly: true }} />
+            <TextField className='output-box' label="Hexadecimal" InputProps={{ readOnly: true }} />
+            <TextField className='output-box' label="Octal" InputProps={{ readOnly: true }} />
+
+        </div>
     )
 }
 
-export default selectBox;
+export default SelectBox;
