+++ src/components/styles/inputbox.css
@@ -10,24 +10,37 @@
     flex-direction: row;
     justify-content: center;
     gap: 10px;
+    margin-bottom: 15px;
+
 
 }
 
 .input-box {
     background: #F4F4F4;
-    height: 40px;
-    width: 30%;
+    border: inset 1px #F4F4F4;
+    height: 50px;
+    width: 40%;
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
 
 }
 
 .select-box {
     background: #528B8E;
-    border: none;
-    height: 40px;
-    width: 15%;
+    border: none;  
+    height: 50px;
+    width: 20%;
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
     text-align: center;
 
+
+}
+
+.submit-box {
+    background: #528B8E;
+    border: none;  
+    height: 50px;
+    width: 20%;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    text-align: center;
 
 }
